refactor(dashboard): apply container/presentational pattern
